BE/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
openai==1.58.1
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.20
